feat: implementa lógica de help global e reorganiza o fluxo de logica

+++ task-manager/src/scripts/register.js
@@ -1,4 +1,4 @@
-const msgFeedback = document.querySelector("#msg-feedback");
+import { showMessage } from "./global.js";
 
 document.querySelector("form").addEventListener("submit", function(event){
     event.preventDefault();
@@ -6,41 +6,37 @@ document.querySelector("form").addEventListener("submit", function(event){
     const username = document.querySelector("input[type='text']").value.trim();
     const email = document.querySelector("input[type='email']").value.trim();
     const password = document.querySelector("input[type='password']").value.trim();
-    
-    // Função para exibir mensagens de feedback/alertas
-    function msgFeed(text , type){
-        msgFeedback.innerText = text;
 
-        msgFeedback.className = "p-3 rounded-md text-sm font-medium mb-4 text-center";
-
-        if (type === "error"){
-            msgFeedback.classList.add("bg-red-100", "text-red-700");
-            msgFeedback.classList.remove("hidden");
-        }
-        else {
-            msgFeedback.classList.add("bg-green-100", "text-green-700", "block");
-            msgFeedback.classList.remove("hidden");
-        }
-    }
-
+    const userList = JSON.parse(localStorage.getItem("users")) || [];
 
     if (!username || !email || !password) {
-        msgFeed("Preencha todos os campos!", "error");
-    return;
+        showMessage("#msg-feedback", "Preencha todos os campos!", "error");
+        return;
     }
 
-const newUser = {
-    username, email , password
-}
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
 
-const userList = JSON.parse(localStorage.getItem("users")) || [];
-userList.push(newUser);
+    if (!emailRegex.test(email)) {
+        showMessage("#msg-feedback", "Por favor, insira um e-mail válido!", "error");
+        return;
+    }
 
-localStorage.setItem("users", JSON.stringify(userList));
-msgFeed("Sucesso! Criando sua conta...", "success");
+    const userExists = userList.some(user => user.email === email);
 
-setTimeout(() => {
-    window.location.href = "login.html";
-  }, 2000);
+    if (userExists) {
+        showMessage("#msg-feedback", "Este e-mail já está cadastrado!", "error");
+        return;
+    }
 
-})
+    const newUser = { username, email, password };
+
+    userList.push(newUser);
+
+    localStorage.setItem("users", JSON.stringify(userList));
+
+    showMessage("#msg-feedback", "Sucesso! Criando sua conta...", "success");
+
+    setTimeout(() => {
+        window.location.href = "login.html";
+    }, 2000);
+});
